@@ -24,62 +24,77 @@
     ]
   },
   "dependencies": {
-    "@fortawesome/fontawesome-svg-core": "^1.3.0",
-    "@fortawesome/free-brands-svg-icons": "^5.15.4",
-    "@fortawesome/free-regular-svg-icons": "^5.15.4",
-    "@fortawesome/free-solid-svg-icons": "^5.15.4",
-    "@fortawesome/react-fontawesome": "^0.1.17",
+    "@fortawesome/fontawesome-svg-core": "^6.1.1",
+    "@fortawesome/free-brands-svg-icons": "^6.1.1",
+    "@fortawesome/free-regular-svg-icons": "^6.1.1",
+    "@fortawesome/free-solid-svg-icons": "^6.1.1",
+    "@fortawesome/react-fontawesome": "^0.1.18",
     "animate.css": "^4.1.1",
-    "nuka-carousel": "^4.8.4",
-    "react": "^17.0.2",
-    "react-countup": "^4.4.0",
-    "react-dom": "^17.0.2",
+    "nuka-carousel": "^5.0.16",
+    "react": "^18.1.0",
+    "react-countup": "^6.2.0",
+    "react-dom": "^18.1.0",
     "react-gtm-module": "^2.0.11",
-    "react-helmet-async": "^1.2.3",
-    "react-router-dom": "^5.3.0",
+    "react-helmet-async": "^1.3.0",
+    "react-router": "^6.3.0",
+    "react-router-dom": "^6.3.0",
     "react-router-hash-link": "^2.4.3",
-    "react-scripts": "^4.0.3",
+    "react-scripts": "^5.0.1",
     "react-scrollspy": "^3.4.3",
     "react-transition-group": "^4.4.2",
     "react-typed": "^1.2.0",
     "react-waypoint": "^10.1.0"
   },
   "devDependencies": {
-    "@types/jest": "^27.4.1",
-    "@types/node": "^17.0.21",
+    "@babel/core": "^7.18.2",
+    "@babel/plugin-syntax-flow": "^7.17.12",
+    "@babel/plugin-transform-react-jsx": "^7.17.12",
+    "@types/jest": "^27.5.1",
+    "@types/node": "^17.0.35",
     "@types/node-sass": "^4.11.2",
-    "@types/react": "^17.0.40",
+    "@types/react": "^18.0.9",
     "@types/react-count-to": "^0.12.1",
-    "@types/react-dom": "^17.0.13",
+    "@types/react-dom": "^18.0.5",
     "@types/react-gtm-module": "^2.0.1",
     "@types/react-router-dom": "^5.3.3",
-    "@types/react-router-hash-link": "^1.2.1",
+    "@types/react-router-hash-link": "^2.4.5",
     "@types/react-scrollspy": "^3.3.5",
     "@types/react-transition-group": "^4.4.4",
-    "@typescript-eslint/eslint-plugin": "^5.14.0",
+    "@typescript-eslint/eslint-plugin": "^5.26.0",
+    "@typescript-eslint/parser": "^5.26.0",
     "babel-eslint": "^10.1.0",
-    "eslint": "^8.11.0",
-    "eslint-config-react-app": "^7.0.0",
+    "eslint": "^8.16.0",
+    "eslint-config-react-app": "^7.0.1",
     "eslint-config-semistandard": "^16.0.0",
-    "eslint-config-standard": "^16.0.3",
+    "eslint-config-standard": "^17.0.0",
     "eslint-plugin-flowtype": "^8.0.3",
-    "eslint-plugin-import": "^2.25.4",
+    "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-jsx-a11y": "^6.5.1",
+    "eslint-plugin-n": "^15.2.0",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-react": "^7.29.3",
-    "eslint-plugin-react-hooks": "^4.3.0",
+    "eslint-plugin-promise": "^6.0.0",
+    "eslint-plugin-react": "^7.30.0",
+    "eslint-plugin-react-hooks": "^4.5.0",
     "node-sass": "^7.0.1",
     "npm-run-all": "^4.1.5",
-    "prettier": "^2.5.1",
-    "prettier-eslint": "^13.0.0",
+    "prettier": "^2.6.2",
+    "prettier-eslint": "^15.0.0",
     "prettier-standard": "^16.4.1",
     "prettier-stylelint": "^0.4.2",
     "semistandard": "^16.0.1",
-    "stylelint": "^14.5.3",
+    "stylelint": "^14.8.5",
     "stylelint-config-prettier": "^9.0.3",
     "stylelint-config-rational-order": "^0.1.2",
     "stylelint-order": "^5.0.0",
     "stylelint-prettier": "^2.0.0",
-    "typescript": "^4.6.2"
+    "typescript": "^4.7.2"
+  },
+  "peerDependencies": {
+    "@babel/core": "^7.18.2",
+    "@babel/plugin-syntax-flow": "^7.17.12",
+    "@babel/plugin-transform-react-jsx": "^7.17.12",
+    "@typescript-eslint/parser": "^5.26.0",
+    "eslint-plugin-n": "^15.2.0",
+    "eslint-plugin-promise": "^6.0.0"
   }
 }
