@@ -1,10 +1,9 @@
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import TagManager from 'react-gtm-module';
 import { HelmetProvider } from 'react-helmet-async';
-
+import App from './components/App';
 import * as serviceWorker from './serviceWorker';
 import './styles/css/index.css';
-import App from './components/App';
 
 TagManager.initialize({
   gtmId: process.env.REACT_APP_GTM_ID!,
@@ -12,11 +11,12 @@ TagManager.initialize({
   preview: process.env.REACT_APP_GTM_PREVIEW,
 });
 
-ReactDOM.render(
+const container = document.getElementById('root');
+const root = createRoot(container!);
+root.render(
   <HelmetProvider>
     <App />
-  </HelmetProvider>,
-  document.getElementById('root')
+  </HelmetProvider>
 );
 
 serviceWorker.register();
